Maksumuistutus käytössä
- puuttuu: Laskujen poisto
- testaamatta: Sähköpostilla lähetys
- listasta ei näe, että maksumuistutus lähetetty

+++ kitupiikki/laskutus/laskudialogi.h
@@ -46,7 +46,7 @@ public:
     ~LaskuDialogi();
 
 private slots:
-    void paivitaSumma(int paivitaSumma);
+    void paivitaSumma(qlonglong paivitaSumma);
     void esikatsele();
     void perusteVaihtuu();
     void haeOsoite();
